[TASK] Rename "artifact migration version" to "artifact migration"
