+++ config/services.php
@@ -43,8 +43,8 @@ return static function (
     DependencyInjection\Loader\Configurator\ContainerConfigurator $configurator,
     DependencyInjection\ContainerBuilder $container,
 ): void {
-    $container->registerForAutoconfiguration(Builder\Artifact\Migration\Version::class)
-        ->addTag('artifact.version_migration')
+    $container->registerForAutoconfiguration(Builder\Artifact\Migration\Migration::class)
+        ->addTag('artifact.migration')
     ;
     $container->registerForAutoconfiguration(Builder\Writer\WriterInterface::class)
         ->addTag('builder.writer')
