+++ tests/config/services.yaml
@@ -17,5 +17,5 @@ services:
   CPSIT\ProjectBuilder\Tests\Fixtures\DummyTemplateRenderingEventListener:
     tags: ['event.listener']
 
-  CPSIT\ProjectBuilder\Tests\Fixtures\DummyVersion:
-    tags: ['artifact.version_migration']
+  CPSIT\ProjectBuilder\Tests\Fixtures\DummyMigration:
+    tags: ['artifact.migration']
